Move NMS packages to cleanup the project
They get relocated to the original location to not alter the resulting jar too much
