+++ modules/betterchairs-plugin/src/main/java/net/blackscarx/betterchairs/ChairsPlugin.java
@@ -2,16 +2,16 @@ package net.blackscarx.betterchairs;
 
 import net.blackscarx.betterchairs.Files.Config;
 import net.blackscarx.betterchairs.Files.Messages;
-import net.blackscarx.betterchairs.NMSManager.NMS_V1_10_R1.v1_10_R1;
-import net.blackscarx.betterchairs.NMSManager.NMS_V1_11_R1.v1_11_R1;
-import net.blackscarx.betterchairs.NMSManager.NMS_V1_13_R2.v1_13_R2;
-import net.blackscarx.betterchairs.NMSManager.NMS_V1_14_R1.v1_14_R1;
-import net.blackscarx.betterchairs.NMSManager.NMS_V1_15_R1.v1_15_R1;
-import net.blackscarx.betterchairs.NMSManager.NMS_v1_8_R1.v1_8_R1;
-import net.blackscarx.betterchairs.NMSManager.NMS_v1_8_R2.v1_8_R2;
-import net.blackscarx.betterchairs.NMSManager.NMS_v1_8_R3.v1_8_R3;
-import net.blackscarx.betterchairs.NMSManager.NMS_v1_9_R1.v1_9_R1;
-import net.blackscarx.betterchairs.NMSManager.NMS_v1_9_R2.v1_9_R2;
+import nms.v1_10_R1.v1_10_R1;
+import nms.v1_11_R1.v1_11_R1;
+import nms.v1_13_R2.v1_13_R2;
+import nms.v1_14_R1.v1_14_R1;
+import nms.v1_15_R1.v1_15_R1;
+import nms.v1_8_R1.v1_8_R1;
+import nms.v1_8_R2.v1_8_R2;
+import nms.v1_8_R3.v1_8_R3;
+import nms.v1_9_R1.v1_9_R1;
+import nms.v1_9_R2.v1_9_R2;
 import net.blackscarx.betterchairs.events.PlayerEnteringChairEvent;
 import net.blackscarx.betterchairs.events.PlayerLeavingChairEvent;
 import net.blackscarx.betterchairs.xseries.XMaterial;
